fixed:#511 search not working fixed (#512) * fixed:#511 search not working fixed
---------

Co-authored-by: ritik kumar <sinharitik589>

## Description

Please provide a brief description of the changes or additions made in this pull request.

## Related Issues

If this pull request is related to any issue(s), please list them here.

## Steps to Test

Provide steps on how to test the changes introduced in this pull request.

## Screenshots (if applicable)

If your changes include visual updates, it would be helpful to provide screenshots of the before and after.

## Checklist

- [ ] I have tested these changes
- [ ] I have updated the relevant documentation
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the codebase
- [ ] My changes generate no new warnings or errors
- [ ] The title of my pull request is clear and descriptive